Slightly fancier compile.sh

+++ compile.sh
@@ -1,2 +1,8 @@
-rm -rf build
+if [ -d "build" ]; then
+  echo "Removing build folder"
+  rm -rf build
+else
+  echo "Build folder does not exist, ignoring..."
+fi
+echo "Running gradle build..."
 gradle build
